requirements.txt
flask
faiss-cpu
numpy
pandas